requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Migrate
flask_cors
selenium
webdriver_manager
pyinstaller
gunicorn
waitress
art
colorama
fake_useragent